Many "unsigned char" to "uint8_t" and "int" to "size_t", so 	that many things compile with -Werror

+++ src/lib/filters.c
@@ -1141,9 +1141,9 @@ ascend_parse_filter(VALUE_PAIR *pair)
  *	Note we don't bother checking 'len' after the snprintf's.
  *	This function should ONLY be called with a large (~1k) buffer.
  */
-void print_abinary(VALUE_PAIR *vp, char *buffer, int len)
+void print_abinary(VALUE_PAIR *vp, char *buffer, size_t len)
 {
-  int 			i;
+  size_t 		i;
   char			*p;
   ascend_filter_t	*filter;
